alert_server/requirements.txt
fastapi
uvicorn
pydantic
pymongo[srv]
python-dateutil
